third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-dynamic-width-expected.txt

Test that creating a cached image and immediately checking its width works

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS currentResult is true
PASS currentResult is true
PASS successfullyParsed is true

TEST COMPLETE
